refactor(memory): extract XxHash64 to kernel codec and mark index tech debt (#428)
Cleanup phase (TD-12/TD-14/TD-15):

- Extract XxHash64 inner class from TextAppendMemory to
  kernel/codec/XxHash64 as standalone utility class
- Confirm Default*Memory classes (DefaultRecordMemory,
  DefaultAppendMemory, DefaultRegistryMemory) are actively used —
  no action needed (TD-14 resolved)
- Add TODO(#428) to MemoryBM25Index and MemorySpladeIndex
  for future common infrastructure extraction

Refs #428
