+++ memory/spector-memory/src/main/java/com/spectrayan/spector/memory/cortex/MemorySpladeIndex.java
@@ -27,6 +27,7 @@ import java.util.Map;
 import java.util.concurrent.Callable;
 import java.util.concurrent.CopyOnWriteArrayList;
 
+// TODO(#428): Extract common index infrastructure to AbstractMemoryIndex
 /**
  * Per-partition SPLADE index manager for learned sparse retrieval.
  *
